organize deployAndWriteGenesis
