+++ scripts/deploy.ts
@@ -1,12 +1,9 @@
 import { ethers } from "ethers";
-import { allContracts } from "../ts/allContractsInterfaces";
-import { deployAll } from "../ts/deploy";
+import { deployAndWriteGenesis } from "../ts/deploy";
 import { DeploymentParameters } from "../ts/interfaces";
 import fs from "fs";
 import { PRODUCTION_PARAMS } from "../ts/constants";
 import { StateTree } from "../ts/stateTree";
-import { execSync } from "child_process";
-import { deployKeyless } from "../ts/deployment/deploy";
 
 const argv = require("minimist")(process.argv.slice(2), {
     string: ["url", "root", "key", "input", "output"]
@@ -44,31 +41,9 @@ async function main() {
 
     parameters.GENESIS_STATE_ROOT =
         argv.root || getDefaultGenesisRoot(parameters);
-
-    const genesisPath = argv.output ?? "genesis.json";
-
     console.log("Deploy with parameters", parameters);
 
-    const genesisEth1Block = await provider.getBlockNumber();
-    await deployKeyless(signer, true);
-    const contracts = await deployAll(signer, parameters, true);
-    let addresses: { [key: string]: string } = {};
-    Object.keys(contracts).map((contract: string) => {
-        addresses[contract] = contracts[contract as keyof allContracts].address;
-    });
-    const appID = await contracts.rollup.appID();
-    const version = execSync("git rev-parse HEAD")
-        .toString()
-        .trim();
-    const auxiliary = {
-        domain: appID,
-        genesisEth1Block,
-        version
-    };
-    const configs = { parameters, addresses, auxiliary };
-    console.log("Writing genesis file to", genesisPath);
-    fs.writeFileSync(genesisPath, JSON.stringify(configs, null, 4));
-    console.log("Successsfully deployed", configs);
+    await deployAndWriteGenesis(signer, parameters, argv.output);
 }
 
 main();
