+++ ts/deploy.ts
@@ -22,6 +22,9 @@ import {
 import { BurnAuctionFactory } from "../types/ethers-contracts/BurnAuctionFactory";
 import { ProofOfBurnFactory } from "../types/ethers-contracts/ProofOfBurnFactory";
 import { GenesisNotSpecified } from "./exceptions";
+import { deployKeyless } from "./deployment/deploy";
+import { execSync } from "child_process";
+import fs from "fs";
 
 async function waitAndRegister(
     contract: Contract,
@@ -169,4 +172,32 @@ export async function deployAll(
 
 async function waitUntilMined(tx: Promise<ContractTransaction>) {
     await (await tx).wait();
+}
+
+export async function deployAndWriteGenesis(
+    signer: Signer,
+    parameters: DeploymentParameters,
+    genesisPath: string = "genesis.json"
+) {
+    let addresses: { [key: string]: string } = {};
+    const genesisEth1Block = await signer.provider?.getBlockNumber();
+    await deployKeyless(signer, true);
+    const contracts = await deployAll(signer, parameters, true);
+
+    Object.keys(contracts).map((contract: string) => {
+        addresses[contract] = contracts[contract as keyof allContracts].address;
+    });
+    const appID = await contracts.rollup.appID();
+    const version = execSync("git rev-parse HEAD")
+        .toString()
+        .trim();
+    const auxiliary = {
+        domain: appID,
+        genesisEth1Block,
+        version
+    };
+    const configs = { parameters, addresses, auxiliary };
+    console.log("Writing genesis file to", genesisPath);
+    fs.writeFileSync(genesisPath, JSON.stringify(configs, null, 4));
+    console.log("Successsfully deployed", configs);
 }
